Move `decode_textures` to global `texture` module
